model/requirements_fasttext.txt
astroid==2.2.5
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
cloudpickle==1.2.1
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
fasttext==0.9.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
idna==2.8
imagesize==1.1.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.3.1
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==19.0.2
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbformat==4.4.0
nltk==3.4.4
numpy==1.16.4
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.9.0
pybind11==2.3.0
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.3
python-dateutil==2.8.0
pytz==2019.1
pywin32-ctypes==0.2.0
pyzmq==18.0.2
requests==2.22.0
scipy==1.3.0
six==1.12.0
snowballstemmer==1.9.0
spyder-kernels==0.5.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
wrapt==1.11.2
